Optimization changes till projektet i metodiken.

+++ Pmetodik/RaylibProject/RaylibProject/element.cpp
@@ -1,5 +1,5 @@
 #include "element.h"
 
-Element::Element(int width, int height, float velocity , Color color) : width(width), height(height), velocity(velocity), color(color)
+Element::Element(int width, int height, int velocity , Color color) : width(width), height(height), velocity(velocity), color(color)
 {}
