+++ Pmetodik/RaylibProject/RaylibProject/element.h
@@ -7,18 +7,18 @@ class Element {
 protected:
 	int width;
 	int height;
-	float velocity;
+	int velocity;
 	Color color;
 
 
 public:
-	Element(int width, int height, float velocity, Color color);
+	Element(int width, int height, int velocity, Color color);
 	virtual std::string getType() = 0;
 	virtual void drawElement() const = 0;
 	virtual int getWidth() const = 0;
 	virtual int getHeight() const = 0;
 	virtual void Gravity() = 0;
-	virtual void setPosition(float x, float y) = 0;
+	virtual void setPosition(int x, int y) = 0;
 	virtual void dontGoBelowFloor(int floorY) = 0;
